re-arranged rob's drum session
took a pass at rearranging Rob's drums that he recorded last time. Found this beat I liked from his track and tried to loop it but grabed the wrong bit and ended up creating a shuffle beat that was pretty cool. Then eqed the hell out of it. Also created an uneqed track for a subtle part on a breakdown. Also kept some of the fills he did on the track for accents. Even Muted the original programed drums part