Tom mixed robs drums
Tom mixed the drums Rob recorded. Ended up using the two last tracks that we did using the 1 mic method and electronic on the other. Fixed some of the parts where the beat would drop. Added a compressor push to the 1 mic drums to give it a bit more edge and equed it with a preamp. Then adjusted the volume so it would sit. Muted the tamborine on the take. Cant decide if I still want it in there or not. Without it's more sparse kinda cool but with it it moves.