Another mix for guitar
Turned the guitars up also slight eqing to it now that they're louder.
Panned the guitar pluck parts in the verse hard left and right and took out the marimba when they played on that cuz they distracted those guitar's back and forth rhythm.
Also added a room reverb to the warble low synth part to make it explode when it hits more.